+++ backend/app/models/ga.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from sqlalchemy import String, Numeric, Text
 from sqlalchemy.orm import Mapped, mapped_column
 from app.models.base import Base, EntityMixin
@@ -8,5 +9,5 @@ class GaEntry(Base, EntityMixin):
 
     category: Mapped[str] = mapped_column(String(100), nullable=False)
     date: Mapped[str] = mapped_column(String(10), nullable=False)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     notes: Mapped[str] = mapped_column(Text, default="")
